third_party/blink/web_tests/fast/scrolling/overflow-auto-ltr-expected.txt

Check that the scroll offset in an LTR block is correctly preserved when the existence of its vertical scrollbar changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS finalOffset is initialOffset
PASS successfullyParsed is true

TEST COMPLETE

INNER INNER INNER
